fix: move billing activity section into admin route scope
