+++ src/cook-web/src/app/admin/billing/AdminBillingPageClient.tsx
@@ -7,7 +7,7 @@ import { useEnvStore } from '@/c-store';
 import { Tabs, TabsContent, TabsList, TabsTrigger } from '@/components/ui/Tabs';
 import { BillingCreditDetailsPanel } from '@/components/billing/BillingCreditDetailsPanel';
 import { BillingOverviewTab } from '@/components/billing/BillingOverviewTab';
-import { BillingRecentActivitySection } from '@/components/billing/BillingRecentActivitySection';
+import { BillingRecentActivitySection } from './components/BillingRecentActivitySection';
 import AdminBreadcrumb from '@/app/admin/components/AdminBreadcrumb';
 import AdminTitle, {
   ADMIN_TITLE_HEADLINE_TABS_LIST_CLASSNAME,
